% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neotoma2LiPD.R
\name{createNeoTsid}
\alias{createNeoTsid}
\title{create a TSid for a Neotoma Lipd}
\usage{
createNeoTsid()
}
\value{
a randomly generated TSid for this project
}
\description{
create a TSid for a Neotoma Lipd
}
