% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeotomaDoi.R
\name{get_neotoma_doi}
\alias{get_neotoma_doi}
\title{Get neotoma doi from dataset
get neotoma doi from dataset name.}
\usage{
get_neotoma_doi(dataset)
}
\arguments{
\item{datasetId}{a list returned from get_neotoma_json or a dataset id}
}
\value{
Neotoma DOI
}
\description{
Get neotoma doi from dataset
get neotoma doi from dataset name.
}
\examples{
get_neotoma_doi(12)
}
